requirements.txt
Flask
werkzeug
razorpay
python-docx
pdfplumber
comtypes
python-dotenv
